+++ README.md
@@ -1,12 +1,13 @@
-# Anpan
-Anpan is a WIP command handler for DiscordGo, inspired by [Clinet](https://github.com/JoshuaDoes/clinet) and [harmony](https://github.com/superwhiskers/harmony)
-
-# Status
 [![Go Report Card](https://goreportcard.com/badge/github.com/MikeModder/anpan)](https://goreportcard.com/report/github.com/MikeModder/anpan)
 [![Build Status](https://travis-ci.org/MikeModder/anpan.svg?branch=master)](https://travis-ci.org/MikeModder/anpan)
 [![GoDoc](https://godoc.org/github.com/MikeModder/anpan?status.svg)](https://godoc.org/github.com/MikeModder/anpan)
 
+# anpan
+**anpan** is a command handler for **[discordgo](https://github.com/bwmarrin/discordgo)**.<br>
+anpan is inspired by **[Clinet](https://github.com/JoshuaDoes/clinet)** and **[harmony](https://github.com/superwhiskers/harmony)**.
+
 # Examples
 Some examples of anpan being used:
-* `examples/` in this repo
-* My bot [007Bot-Go](https://github.com/MikeModder/007Bot-Go)
+* The example bot inside the `examples` directory.
+* My bot, **[007Bot-Go](https://github.com/MikeModder/007Bot-Go)**.
+* Apfel's current bot, **[Nutty Nook](https://github.com/Apfel/Nutty-Nook)**.
