--- command-handler.go
@@ -1,395 +1,0 @@
-package anpan
-
-/* command-handler.go:
- * Contains the main code of the command handler.
- *
- * Anpan (c) 2019 MikeModder/MikeModder007
- */
-
-import (
-	"fmt"
-	"strings"
-
-	"github.com/bwmarrin/discordgo"
-)
-
-// AddPrefix adds a single prefix to the prefixes.
-func (c *CommandHandler) AddPrefix(prefix string) {
-	c.Prefixes = append(c.Prefixes, prefix)
-}
-
-// RemovePrefix removes a single prefix from the prefixes.
-func (c *CommandHandler) RemovePrefix(prefix string) {
-	for i, v := range c.Prefixes {
-		if v == prefix {
-			c.Prefixes = append(c.Prefixes[:i], c.Prefixes[i+1:]...)
-			break
-		}
-	}
-}
-
-// SetPrefixes changes all prefixes.
-func (c *CommandHandler) SetPrefixes(prefixes []string) {
-	c.Prefixes = prefixes
-}
-
-// GetPrefixes gets the current prefixes.
-func (c *CommandHandler) GetPrefixes() []string {
-	return c.Prefixes
-}
-
-// SetPrerunFunc sets the function to run before the command handler's OnMessage.
-func (c *CommandHandler) SetPrerunFunc(prf PrerunFunc) {
-	c.PrerunFunc = prf
-}
-
-// ClearPrerunFunc removes the prerun function.
-func (c *CommandHandler) ClearPrerunFunc() {
-	c.PrerunFunc = nil
-}
-
-// SetOnErrorFunc sets the function to run when a command returns an error
-func (c *CommandHandler) SetOnErrorFunc(oef OnErrorFunc) {
-	c.OnErrorFunc = oef
-}
-
-// ClearOnErrorFunc clears the onerror function
-func (c *CommandHandler) ClearOnErrorFunc() {
-	c.OnErrorFunc = nil
-}
-
-// AddCommand adds a command to the Commands map.
-func (c *CommandHandler) AddCommand(name, desc string, owneronly, hidden bool, perms Permission, cmdtype CommandType, run CommandRunFunc) {
-	c.Commands[name] = &Command{
-		Name:        name,
-		Description: desc,
-		OwnerOnly:   owneronly,
-		Hidden:      hidden,
-		Permissions: perms,
-		Type:        cmdtype,
-		Run:         run,
-	}
-}
-
-// RemoveCommand removes a command from the Commands map.
-func (c *CommandHandler) RemoveCommand(name string) {
-	if _, has := c.Commands[name]; has {
-		delete(c.Commands, name)
-	}
-	return
-}
-
-// IsOwner checks if the given user ID is one of the owners.
-func (c *CommandHandler) IsOwner(id string) bool {
-	for _, o := range c.Owners {
-		if id == o {
-			return true
-		}
-	}
-
-	return false
-}
-
-// SetDebug sets debug on or off.
-func (c *CommandHandler) SetDebug(enabled bool) {
-	c.Debug = enabled
-}
-
-func (c *CommandHandler) debugLog(out string) {
-	if c.Debug {
-		fmt.Println(out)
-	}
-}
-
-// AddDefaultHelpCommand adds the default (library provided) help command to the list of commands.
-// TODO: users have to manually call this to add the help command, maybe find a way to add it automatially if no help command is detected?
-func (c *CommandHandler) AddDefaultHelpCommand() {
-	c.AddCommand("help", "Get some help about using the bot.", false, false, 0, CommandTypeEverywhere, c.defaultHelpCmd)
-}
-
-// OnMessage - You don't need to call this! Pass this to AddHandler().
-func (c *CommandHandler) OnMessage(s *discordgo.Session, m *discordgo.MessageCreate) {
-	// Check if the author is a bot, and deny entry if IgnoreBots is true.
-	if m.Author.Bot && c.IgnoreBots {
-		c.debugLog("Author is bot")
-		return
-	}
-
-	content := m.Content
-	c.debugLog(content)
-
-	// Check for one of the prefixes. If the content doesn't start with one of the prefixes, return.
-	var prefix string
-	var has bool // false by default
-	for i := 0; i < len(c.Prefixes); i++ {
-		prefix = c.Prefixes[i]
-		if strings.HasPrefix(content, prefix) {
-			has = true
-			break
-		}
-	}
-
-	// Check is has is true. If it's not that means we didn't find a prefix and should return
-	if !has {
-		return
-	}
-
-	// Since we just checked for a prefix, now we need to trim off the prefix
-	cmd := strings.Split(strings.TrimPrefix(content, prefix), " ")
-	c.debugLog(cmd[0])
-
-	// Check and see if we have a command by that name.
-	if command, exist := c.Commands[cmd[0]]; exist {
-		// We do, so check permissions.
-		// We do, so check if this channel is Private Messages; they don't support permissions.
-		channel, err := s.Channel(m.ChannelID)
-		if err != nil {
-			c.debugLog("Couldn't retrieve Channel, continuing...")
-		}
-
-		var dm bool // again, false by default.
-		if channel.Type == discordgo.ChannelTypeDM {
-			dm = true
-		}
-
-		// If it isn't Private Messages, go ahead and check the permissions.
-		if command.Type != CommandTypePrivate && !dm && !checkPermissions(s, m.GuildID, m.Author.ID, command.Permissions) {
-			embed := &discordgo.MessageEmbed{
-				Title:       "Insufficient Permissions!",
-				Description: "You don't have the correct permissions to run this command!",
-				Color:       0xff0000,
-			}
-
-			if !command.Hidden {
-				s.ChannelMessageSendEmbed(m.ChannelID, embed)
-			}
-
-			c.debugLog("Insufficient permissions for User.")
-			return
-		}
-
-		// Same here, just for the bot itself.
-		if command.Type != CommandTypePrivate && !dm && !checkPermissions(s, m.GuildID, s.State.User.ID, command.Permissions) {
-			embed := &discordgo.MessageEmbed{
-				Title:       "Insufficient Permissions!",
-				Description: "Uh oh, I don't have the right permissions to execute that command for you! Make sure I have the right permissions (ex. Kick Members) then try running the command again!",
-				Color:       0xff0000,
-			}
-
-			if !command.Hidden {
-				s.ChannelMessageSendEmbed(m.ChannelID, embed)
-			}
-
-			c.debugLog("Insufficient permissions for Bot.")
-			return
-		}
-
-		if channel.Type == discordgo.ChannelTypeDM && command.Type == CommandTypeGuild {
-			embed := &discordgo.MessageEmbed{
-				Title:       "Invalid Channel!",
-				Description: "You cannot run this command on a guild!",
-				Color:       0xff0000,
-			}
-
-			if !command.Hidden {
-				s.ChannelMessageSendEmbed(m.ChannelID, embed)
-			}
-
-			c.debugLog("Tried to run DM-only command on a guild.")
-			return
-		} else if channel.Type == discordgo.ChannelTypeGuildText && command.Type == CommandTypePrivate {
-			embed := &discordgo.MessageEmbed{
-				Title:       "Invalid Channel!",
-				Description: "You cannot run this command inside Private Messages!",
-				Color:       0xff0000,
-			}
-
-			if !command.Hidden {
-				s.ChannelMessageSendEmbed(m.ChannelID, embed)
-			}
-
-			c.debugLog("Tried to run Guild-Only command inside DM.")
-			return
-		}
-
-		// Check if it's an owner-only command, and if it is make sure the author is an owner.
-		if command.OwnerOnly && !c.IsOwner(m.Author.ID) {
-			// It's an owner-only command, and the user wasn't on the owners list.
-			embed := &discordgo.MessageEmbed{
-				Title:       "You can't run that command!",
-				Description: "Sorry, only the bot owner(s) can run that command!",
-				Color:       0xff0000,
-			}
-
-			// If command isn't hidden, send a message.
-			if !command.Hidden {
-				s.ChannelMessageSendEmbed(m.ChannelID, embed)
-			}
-			c.debugLog("Owner only command.")
-			return
-		}
-
-		c.debugLog(fmt.Sprintf("Executing %s.", command.Name))
-
-		c.debugLog("Firing PrerunFunc")
-		if c.PrerunFunc != nil {
-			// Run the user's prerun function
-			// if it returns true, then stop here.
-			if c.PrerunFunc(s, m, command.Name, cmd[1:]) {
-				return
-			}
-		}
-
-		guild, err := s.Guild(m.GuildID)
-		if err != nil {
-			c.debugLog("Couldn't retrieve Guild, continuing...")
-		}
-
-		member, err := s.State.Member(m.GuildID, m.Author.ID)
-		if err != nil {
-			c.debugLog("Couldn't retrieve Member, continuing...")
-		}
-
-		context := Context{
-			Session: s,
-			Message: m.Message,
-			User:    m.Author,
-			Channel: channel,
-			Guild:   guild,
-			Member:  member,
-		}
-
-		err = command.Run(context, cmd[1:])
-		if err != nil && c.OnErrorFunc != nil {
-			// Run the user's OnErrorFunc
-			c.OnErrorFunc(context, cmd[0], err)
-		}
-
-	} else {
-		// We don't :(
-		c.debugLog("Unknown command / not even one.")
-		return
-	}
-}
-
-func (c *CommandHandler) defaultHelpCmd(context Context, args []string) error {
-	if len(args) >= 1 {
-		if commannd, has := c.Commands[args[0]]; has {
-			// Check if the command is hidden.
-			if commannd.Hidden {
-				return nil
-			}
-
-			// basic channel check
-			if context.Channel.Type == discordgo.ChannelTypeDM && commannd.Type == CommandTypeGuild {
-				return nil
-			} else if context.Channel.Type == discordgo.ChannelTypeGuildText && commannd.Type == CommandTypePrivate {
-				return nil
-			}
-
-			// Proper English is always fun.
-			var (
-				owneronlystring string
-				typestring      string
-			)
-
-			if commannd.OwnerOnly {
-				owneronlystring = "Yes"
-			} else {
-				owneronlystring = "No"
-			}
-
-			if commannd.Type == CommandTypePrivate {
-				typestring = "Private"
-			} else if commannd.Type == CommandTypeGuild {
-				typestring = "Guild-only"
-			} else {
-				typestring = "No restrictions"
-			}
-
-			embed := &discordgo.MessageEmbed{
-				Title: "Help!",
-				Description: fmt.Sprintf("Help for command `%s`\n Description: `%s`\nOwner only: **%s**\nType: **%s**", commannd.Name, commannd.Description,
-					owneronlystring, typestring),
-				Footer: &discordgo.MessageEmbedFooter{
-					Text: fmt.Sprintf("The bot's prefixes are %s.", c.Prefixes),
-				},
-			}
-
-			context.ReplyEmbed(embed)
-			return nil
-		}
-
-		embed := &discordgo.MessageEmbed{
-			Title:       "Error!",
-			Description: fmt.Sprintf("`%s` is not a valid command!", args[0]),
-			Color:       0xff0000,
-		}
-
-		context.ReplyEmbed(embed)
-		return nil
-
-	}
-
-	count := len(c.Commands)
-	var list string
-
-	for name := range c.Commands {
-		cmd := c.Commands[name]
-		if !cmd.Hidden && typeCheck(context.Channel.Type, cmd.Type) {
-			list += fmt.Sprintf("**%s** - `%s`\n", cmd.Name, cmd.Description)
-		}
-	}
-
-	var footer strings.Builder
-
-	//"The bot's prefixes are %s | There are %d commands." fmt.Sprintf(footer, c.Prefixes, count)
-
-	// Grammar is always fun.
-	if count == 1 {
-		footer.WriteString("There is 1 command.")
-	} else {
-		footer.WriteString(fmt.Sprintf("There are %d commands.", count))
-	}
-
-	footer.WriteString(" | ")
-
-	if len(c.Prefixes) == 1 {
-		footer.WriteString(fmt.Sprintf("The bot's prefix is %s.", c.Prefixes[0]))
-	} else {
-		prefixes := strings.Builder{}
-
-		for i, prefix := range c.Prefixes {
-			if i+1 == len(c.Prefixes) {
-				prefixes.WriteString(fmt.Sprintf("and %s.", prefix))
-			} else {
-				prefixes.WriteString(fmt.Sprintf("%s, ", prefix))
-			}
-		}
-
-		footer.WriteString(fmt.Sprintf("The bot's prefixes are %s.", prefixes.String()))
-	}
-
-	embed := &discordgo.MessageEmbed{
-		Title:       "Commands:",
-		Description: list,
-		Footer: &discordgo.MessageEmbedFooter{
-			Text: footer.String(),
-		},
-	}
-
-	context.ReplyEmbed(embed)
-	return nil
-}
-
-func typeCheck(chn discordgo.ChannelType, cmd CommandType) bool {
-	if cmd == CommandTypeEverywhere {
-		return true
-	} else if chn == discordgo.ChannelTypeDM && cmd == CommandTypePrivate {
-		return true
-	} else if chn == discordgo.ChannelTypeGuildText && cmd == CommandTypeGuild {
-		return true
-	} else {
-		return false
-	}
-}
