+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/MikeModder/anpan
 
-require github.com/bwmarrin/discordgo v0.19.0
+go 1.14
+
+require github.com/bwmarrin/discordgo v0.20.2
