+++ types.go
@@ -3,62 +3,76 @@ package anpan
 /* types.go:
  * Declares all types used for and within anpan.
  *
- * Anpan (c) 2019 MikeModder/MikeModder007
+ * anpan (c) 2020 MikeModder/MikeModder007, Apfel
  */
 
 import "github.com/bwmarrin/discordgo"
 
+// CommandRunFunc is a command's run function.
+type CommandRunFunc func(Context, []string) error
+
+// DebugFunc is used for debugging output.
+type DebugFunc func(string)
+
+// HelpRunFunc is used for definitions of a help command.
+type HelpRunFunc func(context Context, args []string, commands []*Command, prefixes []string) error
+
 // PrerunFunc is the type for the function that can be run before command execution.
+// If all goes well, return true. otherwise, false.
 type PrerunFunc func(*discordgo.Session, *discordgo.MessageCreate, string, []string) bool
 
 // OnErrorFunc is the type for the function that can be run.
 type OnErrorFunc func(Context, string, error)
 
-// CommandRunFunc is a command's run function
-type CommandRunFunc func(Context, []string) error
-
 // CommandType defines where commands can be used.
 type CommandType int
 
-// Permission defines a permission with its associated value.
-type Permission int
+// HelpCommand defines a help command.
+type HelpCommand struct {
+	Aliases []string
+	Name    string
+	Show    bool
+	Run     HelpRunFunc
+}
+
+// CommandHandler contains all the data needed for the handler to function.
+type CommandHandler struct {
+	CheckPermissions bool
+	Commands         []*Command
+	DebugFunc        DebugFunc
+	HelpCommand      HelpCommand
+	IgnoreBots       bool
+	OnErrorFunc      OnErrorFunc
+	Owners           []string
+	Prefixes         []string
+	PrerunFunc       PrerunFunc
+}
+
+// Command represents a command.
+type Command struct {
+	Aliases     []string
+	Hidden      bool
+	Description string
+	Name        string
+	OwnerOnly   bool
+	Permissions int
+	Run         CommandRunFunc
+	Type        CommandType
+}
+
+// Context holds the data required for command execution.
+type Context struct {
+	Channel *discordgo.Channel
+	Guild   *discordgo.Guild
+	Member  *discordgo.Member
+	Message *discordgo.Message
+	Session *discordgo.Session
+	User    *discordgo.User
+}
 
 // Command types; Either DM-Only, Guild-Only or both.
 const (
 	CommandTypePrivate CommandType = iota
 	CommandTypeGuild
 	CommandTypeEverywhere
-
-	PermissionAdministrator       Permission = 8
-	PermissionViewAuditLog        Permission = 128
-	PermissionViewServerInsights  Permission = 524288
-	PermissionManageServer        Permission = 32
-	PermissionManageRoles         Permission = 268435456
-	PermissionManageChannels      Permission = 16
-	PermissionKickMembers         Permission = 2
-	PermissionBanMembers          Permission = 4
-	PermissionCreateInstantInvite Permission = 1
-	PermissionChangeNickname      Permission = 67108864
-	PermissionManageNicknames     Permission = 134217728
-	PermissionManageEmojis        Permission = 1073741824
-	PermissionManageWebhooks      Permission = 536870912
-	PermissionViewChannels        Permission = 1024
-
-	PermissionMessagesSend              Permission = 2048
-	PermissionMessagesSendTTS           Permission = 4096
-	PermissionMessagesManage            Permission = 8192
-	PermissionMessagesEmbedLinks        Permission = 16384
-	PermissionMessagesAttachFiles       Permission = 32768
-	PermissionMessagesReadHistory       Permission = 65536
-	PermissionMessagesMentionEveryone   Permission = 131072
-	PermissionMessagesUseExternalEmojis Permission = 262144
-	PermissionMessagesAddReactions      Permission = 64
-
-	PermissionVoiceConnect         Permission = 1048576
-	PermissionVoiceSpeak           Permission = 2097152
-	PermissionVoiceMuteMembers     Permission = 4194304
-	PermissionVoiceDeafenMembers   Permission = 8388608
-	PermissionVoiceUseMembers      Permission = 16777216
-	PermissionVoiceUseActivity     Permission = 33554432
-	PermissionVoicePrioritySpeaker Permission = 256
 )
